Rename pattern library page modules to match their default exports
