+++ src/pattern-library/components/patterns/layout/PanelPage.js
